moving sbe_reader
